+++ lib/renderer_canvas.dart
@@ -1,0 +1,4 @@
+/// The Canvas renderer, this is required for displaying in a web context
+library cellular_automata.renderer.canvas;
+
+export 'package:cellular_automata/src/renderers/canvas.dart';